requirements.txt
streamlit>=1.30.0
openai>=1.0.0
azure-identity>=1.15.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
